Removed unneeded enviro variables
Originally had/was going to have environmental variables that would point to program files and program data on the remote system. Decided that as these locations seem to be the same across environments (in my experience) that default values would work. Kept the temp variable as that DOES change between hosts.